Refactor canvas module into separate files (Phase 1.5) (#33)
Split the monolithic _canvas_impl.py (1098 lines) into three focused modules:
- layer.py: TemporalLayer class and to_temporal_layer function (427 lines)
- compiled.py: CompiledRHGCanvas class and add_temporal_layer function (458 lines)
- canvas.py: RHGCanvas and RHGCanvasSkeleton classes (259 lines)

Benefits:
- Each file now has a single, well-defined responsibility
- Improved code organization and maintainability
- Easier to navigate and understand
- Backward compatibility maintained via __init__.py

All tests pass (111 passed, 2 skipped)
Code quality checks pass (ruff, mypy)

Closes #33 (Phase 1.5)
